tests/wpt/web-platform-tests/content-security-policy/style-src/style-src-3_3.html

<!DOCTYPE HTML>
<html>
<head>
    <title>href of link with rel=stylesheet must be in src list</title>
    <script src='/resources/testharness.js'></script>
    <script src='/resources/testharnessreport.js'></script>
    <script>
        var head = document.getElementsByTagName('head')[0];
        var link = document.createElement('link');
        link.setAttribute('rel', 'stylesheet');
        link.setAttribute('type', 'text/css');
        link.setAttribute('href', location.protocol +
                            '//www1.' +
                            location.hostname +
                            ':' +
                            location.port +
                            '/content-security-policy/style-src/3_3.css');
        head.appendChild(link);

       onload = function doTest() {
         test(function() {
           var text = document.getElementById("content");
           assert_true(getComputedStyle(text).marginLeft != "2px", "Style sheet loaded from origin not in style-src directive should be blocked");
         });
       }
    </script>
</head>
<body>
    <h1>href of link with rel=stylesheet must be in src list</h1>
    <div id='log'></div>

    <div id="content">This text should not have a margin-left of 2</div>

    <script async defer src='../support/checkReport.sub.js?reportField=violated-directive&reportValue=style-src%20%27self%27'></script>

</body>
</html>
